fix: refactor validation for add and send money form
