+++ dashboard/app/lib/constants/messages.ts
@@ -26,6 +26,8 @@ export const ERROR_MESSAGES = {
   LIMIT_AMOUNT: 'Price cannot exceed 1 million dollars',
   LIMIT_QUANTITY: 'Quantity must not exceed 1 million',
   QUANTITY_INVALID: 'Quantity is invalid',
+  INVALID_TRANSFER_AMOUNT: 'Amount is invalid',
+  LIMIT_TRANSFER_AMOUNT: 'Amount cannot exceed 1 million dollar',
   SET_PIN_CODE: {
     title: 'Set PIN code failed',
     description:
