+++ dashboard/app/lib/utils/index.ts
@@ -15,3 +15,4 @@ export * from './firebase';
 export * from './recentActivities';
 export * from './logAcitvity';
 export * from './updateColorScheme';
+export * from './validation';
